requirements/test.txt
django-debug-toolbar==4.2.0
parameterized==0.9.0
-r prod.txt